// kestrel/src/types/engine/ComplexityResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Analysis } from "./Analysis";
import type { ComplexityPoint } from "./ComplexityPoint";

export type ComplexityResult = { points: Array<ComplexityPoint>, analysis: Analysis, };